requirements.txt
adafruit-blinka==8.62.0
adafruit-circuitpython-busdevice==5.2.13
adafruit-circuitpython-connectionmanager==3.1.5
adafruit-circuitpython-dht==4.0.9
adafruit-circuitpython-requests==4.1.13
adafruit-circuitpython-typing==1.12.1
adafruit-platformdetect==3.81.0
adafruit-pureio==1.1.11
binho-host-adapter==0.1.6
blinker==1.9.0
boto3==1.39.11
botocore==1.39.11
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
flask==3.1.1
idna==3.10
itsdangerous==2.2.0
jinja2==3.1.6
jmespath==1.0.1
markupsafe==3.0.2
pillow==11.3.0
pyftdi==0.56.0
pyserial==3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pyusb==1.3.1
requests==2.32.4
s3transfer==0.13.1
six==1.17.0
sysv-ipc==1.1.0
typing-extensions==4.14.1
urllib3==2.5.0
waitress==3.0.2
werkzeug==3.1.3